#KA-87 separate user and courier proposal
